Pull out base scrapers
