+++ app/scraper/loot/amazon_games.py
@@ -1,40 +1,21 @@
 import logging
-from dataclasses import dataclass
 from datetime import date, datetime, time, timedelta, timezone
 
-from playwright.async_api import Error, Locator, Page
+from playwright.async_api import Locator, Page
 
-from app.common import OfferDuration, OfferType, Source
+from app.common import OfferType
 from app.scraper.info.utils import clean_game_title
-from app.scraper.loot.scraper import OfferHandler, RawOffer, Scraper
+from app.scraper.loot.amazon_base import AmazonBaseScraper, AmazonRawOffer
+from app.scraper.loot.scraper import OfferHandler, Scraper
 from app.sqlalchemy import Offer
 
 logger = logging.getLogger(__name__)
 
-BASE_URL = "https://gaming.amazon.com"
-OFFER_URL = BASE_URL + "/home"
 
-
-@dataclass
-class AmazonRawOffer(RawOffer):
-    valid_to: str | None = None
-
-
-class AmazonGamesScraper(Scraper):
-    @staticmethod
-    def get_source() -> Source:
-        return Source.AMAZON
-
+class AmazonGamesScraper(AmazonBaseScraper):
     @staticmethod
     def get_type() -> OfferType:
         return OfferType.GAME
-
-    @staticmethod
-    def get_duration() -> OfferDuration:
-        return OfferDuration.CLAIMABLE
-
-    def get_offers_url(self) -> str:
-        return OFFER_URL
 
     def get_page_ready_selector(self) -> str:
         return ".offer-list__content"
@@ -53,57 +34,13 @@ class AmazonGamesScraper(Scraper):
         await Scraper.scroll_element_to_bottom(page, "root")
 
     async def read_raw_offer(self, element: Locator) -> AmazonRawOffer:
-        title = await element.locator(
-            ".item-card-details__body__primary h3"
-        ).text_content()
-        if title is None:
-            raise ValueError("Couldn't find title.")
-
-        valid_to = await element.locator(
-            ".item-card__availability-date p"
-        ).text_content()
-        if valid_to is None:
-            raise ValueError(f"Couldn't find valid to for {title}.")
-
-        img_url = await element.locator(
-            '[data-a-target="card-image"] img'
-        ).get_attribute("src")
-        if img_url is None:
-            raise ValueError(f"Couldn't find image for {title}.")
-
-        url = BASE_URL
-
-        try:
-            path = await element.locator(
-                '[data-a-target="learn-more-card"]'
-            ).get_attribute("href", timeout=500)
-            if path is not None:
-                url += path
-        except Error:
-            # Some offers are claimed on site and don't have a specific path.
-            # That's fine.
-            pass
-
-        return AmazonRawOffer(
-            title=title,
-            valid_to=valid_to,
-            url=url,
-            img_url=img_url,
-        )
+        return await self.read_base_raw_offer(element)
 
     def normalize_offers(self, raw_offers: list[AmazonRawOffer]) -> list[Offer]:  # type: ignore
         normalized_offers: list[Offer] = []
 
         for raw_offer in raw_offers:
-            # Raw text
-            if not raw_offer.title:
-                logger.error(f"Error with offer, has no title: {raw_offer}")
-                continue
-
             rawtext = f"<title>{raw_offer.title}</title>"
-
-            # Title
-            probable_game_name: str | None = None
             probable_game_name = clean_game_title(raw_offer.title)
 
             # Date
