+++ app/scraper/loot/amazon_loot.py
@@ -2,43 +2,25 @@ import logging
 from dataclasses import dataclass
 from datetime import date, datetime, time, timedelta, timezone
 
-from playwright.async_api import Error, Locator, Page
+from playwright.async_api import Locator, Page
 
-from app.common import OfferDuration, OfferType, Source
-from app.scraper.info.utils import clean_loot_title
-from app.scraper.loot.scraper import OfferHandler, RawOffer, Scraper
+from app.common import OfferType
+from app.scraper.loot.amazon_base import AmazonBaseScraper, AmazonRawOffer
+from app.scraper.loot.scraper import OfferHandler, Scraper
 from app.sqlalchemy import Offer
 
 logger = logging.getLogger(__name__)
 
-BASE_URL = "https://gaming.amazon.com"
-OFFER_URL = BASE_URL + "/home"
-
 
 @dataclass
-class AmazonLootRawOffer(RawOffer):
+class AmazonLootRawOffer(AmazonRawOffer):
     game_title: str | None = None
-    valid_to: str | None = None
 
 
-class AmazonLootScraper(Scraper):
-    @staticmethod
-    def get_source() -> Source:
-        return Source.AMAZON
-
+class AmazonLootScraper(AmazonBaseScraper):
     @staticmethod
     def get_type() -> OfferType:
         return OfferType.LOOT
-
-    @staticmethod
-    def get_duration() -> OfferDuration:
-        return OfferDuration.CLAIMABLE
-
-    def get_offers_url(self) -> str:
-        return OFFER_URL
-
-    def get_page_ready_selector(self) -> str:
-        return ".offer-list__content"
 
     def get_offer_handlers(self, page: Page) -> list[OfferHandler]:
         return [
@@ -54,72 +36,29 @@ class AmazonLootScraper(Scraper):
         await Scraper.scroll_element_to_bottom(page, "root")
 
     async def read_raw_offer(self, element: Locator) -> AmazonLootRawOffer:
+        base_raw_offer = await self.read_base_raw_offer(element)
+
         game_title = await element.locator(".item-card-details__body p").text_content()
         if game_title is None:
             raise ValueError("Couldn't find game title.")
 
-        title = await element.locator(
-            ".item-card-details__body__primary h3"
-        ).text_content()
-        if title is None:
-            raise ValueError("Couldn't find title.")
-
-        valid_to = await element.locator(
-            ".item-card__availability-date p"
-        ).text_content()
-        if valid_to is None:
-            raise ValueError(f"Couldn't find valid to for {title}.")
-
-        img_url = await element.locator(
-            '[data-a-target="card-image"] img'
-        ).get_attribute("src")
-        if img_url is None:
-            raise ValueError(f"Couldn't find image for {title}.")
-
-        url = BASE_URL
-
-        try:
-            path = await element.locator(
-                '[data-a-target="learn-more-card"]'
-            ).get_attribute("href", timeout=500)
-            if path is not None:
-                url += path
-        except Error:
-            # Some offers are claimed on site and don't have a specific path.
-            # That's fine.
-            pass
-
         return AmazonLootRawOffer(
-            title=title,
+            title=base_raw_offer.title,
+            valid_to=base_raw_offer.valid_to,
+            url=base_raw_offer.url,
+            img_url=base_raw_offer.img_url,
             game_title=game_title,
-            valid_to=valid_to,
-            url=url,
-            img_url=img_url,
         )
 
     def normalize_offers(self, raw_offers: list[AmazonLootRawOffer]) -> list[Offer]:  # type: ignore
         normalized_offers: list[Offer] = []
 
         for raw_offer in raw_offers:
-            # Raw text
-            if not raw_offer.title:
-                logger.error(f"Error with offer, has no title: {raw_offer}")
-                continue
-
             rawtext = f"<title>{raw_offer.title}</title>"
 
-            # Game title
-            # New Amazon page layout since 2022-08-09 21:10 finally includes
-            # the title in a seperate tag
-            if raw_offer.game_title:
-                rawtext += f"<gametitle>{raw_offer.game_title}</gametitle>"
-                probable_game_name = raw_offer.game_title
-                title = f"{raw_offer.game_title}: {raw_offer.title}"
-            # If the title is not there for any reason, try to get it from the
-            # general loot description instead
-            else:
-                title = raw_offer.title
-                probable_game_name = clean_loot_title(raw_offer.title)
+            rawtext += f"<gametitle>{raw_offer.game_title}</gametitle>"
+            probable_game_name = raw_offer.game_title
+            title = f"{raw_offer.game_title}: {raw_offer.title}"
 
             # Date
             # This is a bit more complicated as only the relative end is
